PHOENIX-6557 Fix code problems flagged by SpotBugs as High priority

+++ phoenix-core/src/main/java/org/apache/phoenix/cache/aggcache/SpillManager.java
@@ -249,7 +249,7 @@ public class SpillManager implements Closeable {
     // Determines the partition, i.e. spillFile the tuple should get spilled to.
     private int getPartition(ImmutableBytesWritable key) {
         // Simple implementation hash mod numFiles
-        return Math.abs(key.hashCode()) % numSpillFiles;
+        return (int)(Math.abs((long)key.hashCode()) % numSpillFiles);
     }
 
     /**
